Implement period button with period check
The checkPeriod function checks the amount of periods for every
number in the number array. If the amount of periods in a number
is greater than one, it stops the function and the calculator
outputs an error.

The regex in the operator array's split method is updated to
account for the period(s) and decimal numbers in the input.